requirements.txt
apache-airflow
flask
kafka-python
cassandra-driver
pyspark